refactor: 파일탐색기 폴더 구조 및 파일명 개선
- constants 폴더 파일명 일관성 개선 (error-messages → error-constants 등)
- lib 폴더 파일명 개선 (drop-zone-utils → drop-zone-helpers)
- 모든 import 경로 업데이트
- 채팅 폴더와 일관성 있는 명명 규칙 적용
