+++ src/features/file-explorer/constants/error-constants.ts
