+++ src/features/file-explorer/lib/drop-zone-helpers.ts
