+++ src/features/file-explorer/ui/file-explorer-panel.tsx
@@ -3,7 +3,7 @@
 import { useParams } from 'next/navigation'
 import { useMemo, useRef } from 'react'
 import { FILE_TREE_CONSTANTS } from '@/features/file-explorer/constants/file-tree-constants'
-import { FILE_EXPLORER_UI_TEXTS } from '@/features/file-explorer/constants/ui-texts'
+import { FILE_EXPLORER_UI_TEXTS } from '@/features/file-explorer/constants/ui-constants'
 import { useFileActions } from '@/features/file-explorer/hooks/use-file-actions'
 import useFileTree from '@/features/file-explorer/hooks/use-file-tree'
 import FileCreateDialog, {
